[ft #161915370] sales endpoint restriction to specific users, admin or attendant

+++ app/views/products.py
@@ -1,3 +1,4 @@
+import re
 from flask import request, jsonify, make_response
 from flask_restful import Resource
 from flask_jwt_extended import (JWTManager, jwt_required, get_jwt_claims)
@@ -33,6 +34,10 @@ class Products(Resource):
         check_product = [product for product in product_exist if product["product_name"]==product_name]
         if check_product:
             return make_response(jsonify({"message":"product already exist"}), 409)
+        elif re.search(r"\s", product_quantity):
+            return make_response(jsonify({"message":"product Quantity cannot be a whitespace"}),404)
+        elif re.search(r"\s", price):
+            return make_response(jsonify({"message":"price cannot be a whitespace"}),404)
         elif not product_category:
             return make_response(jsonify({"message":"Product category required"}),404)
         elif not product_quantity:
@@ -97,6 +102,10 @@ class DeleteProduct(Resource):
 
     @jwt_required
     def delete(self, product_id):
+        claims = get_jwt_claims()
+        if claims['role'] != "admin":
+            return make_response(jsonify({"message": "Sorry, you don't have administrator rights"}), 403)
+            
         delete_product = self.user.delete_product(product_id)
         return delete_product
